+++ src/main/java/com/isoft/airport/models/Flight.java
@@ -1,6 +1,10 @@
 package com.isoft.airport.models;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 import javax.persistence.*;
 import java.time.LocalDate;
 import java.util.HashSet;
@@ -16,7 +20,7 @@ public class Flight extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "flight_id", insertable = false, updatable = false)
-    private int flightId;
+    private long flightId;
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinColumn(name = "from")
     private Airport fromAirport;
