+++ src/main/java/com/isoft/airport/repositories/BookingRepository.java
@@ -3,5 +3,5 @@ package com.isoft.airport.repositories;
 import com.isoft.airport.models.Booking;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface BookingRepository extends PagingAndSortingRepository<Booking,Integer> {
+public interface BookingRepository extends PagingAndSortingRepository<Booking,Long> {
 }
